fix some weird breaking change related to ctx
ctx wasnt working with mems created in code. eg, this was compiled with wat2wasm and executed in pywasm:

```(module
    (import "lib" "write" (func $write (param i32)))
    (memory $m 1)
    (data (i32.const 0) "Hi")
    (func $program
        i32.const 22
        call $write
    )
    (start $program)
)```

where lib.write is a python function accepting two args (ctx,val) just to check ctx is passing properly to HostFuncs. but this was crashing due to some weird changes in hostfunc_call and exec_expr from 0.4.5. just revert these 3 lines or so to fix for our project.